Extract `hex_to_dec_netmask` into mixin/network_helper.rb
